requirements.txt
pystray>=0.19.0
keyboard>=0.13.5
Pillow>=10.0.0
pyinstaller>=6.3.0
customtkinter>=5.2.0
